+++ .eslintrc.yml
@@ -19,5 +19,6 @@ rules: {
   "no-restricted-syntax": off,
   "guard-for-in": off,  
   "no-underscore-dangle": off,
+  "no-alert": off,
  }
