+++ package.json
@@ -10,9 +10,8 @@
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
     "prettier": "2.0.5",
-    "webpack-cli": "^3.3.11",    
-    "webpack": "^4.43.0"    
+    "webpack-cli": "^3.3.11",
+    "webpack": "^4.43.0"
   },
-  "dependencies": {
-  }
+  "dependencies": {}
 }
