+++ src/js/models/filters-model.js
@@ -1,0 +1,7 @@
+class FiltersModel {
+  constructor(filters) {
+    this.filters = filters;
+  }
+}
+
+export { FiltersModel };
